+++ src/main/java/com/example/moviereserve/entity/venues/Venues.java
@@ -1,4 +1,4 @@
-package com.example.moviereserve.entity;
+package com.example.moviereserve.entity.venues;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
