+++ src/main/java/com/example/moviereserve/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.example.moviereserve.repository;
 
-import com.example.moviereserve.entity.User;
+import com.example.moviereserve.entity.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
