+++ src/main/java/com/example/moviereserve/repository/VenuesRepository.java
@@ -1,6 +1,6 @@
 package com.example.moviereserve.repository;
 
-import com.example.moviereserve.entity.Venues;
+import com.example.moviereserve.entity.venues.Venues;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
