+++ src/main/java/com/example/moviereserve/service/VenuesService.java
@@ -2,9 +2,8 @@ package com.example.moviereserve.service;
 
 import com.example.moviereserve.dto.venues.VenuesResponseDto;
 import com.example.moviereserve.dto.venues.VenuesSetupRequestDto;
-import com.example.moviereserve.entity.User;
-import com.example.moviereserve.entity.Venues;
-import com.example.moviereserve.entity.VenuesType;
+import com.example.moviereserve.entity.venues.Venues;
+import com.example.moviereserve.entity.venues.VenuesType;
 import com.example.moviereserve.repository.VenuesRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
